Updated data types.

+++ accelerate-stepper-motor-aryeh-eiderman-04-algorithm/accelerate-stepper-motor-aryeh-eiderman-04-algorithm.ino
@@ -9,9 +9,9 @@
 
 /// @{
 /// @brief GPIO pins.
-const uint8_t kPulPin = 11; ///< For the stepper driver PUL/STP/CLK (pulse/step) pin.
-const uint8_t kDirPin = 12; ///< For the stepper driver DIR/CW (direction) pin.
-const uint8_t kEnaPin = 13; ///< For the stepper driver ENA/EN (enable) pin.
+const uint16_t kPulPin = 11; ///< For the stepper driver PUL/STP/CLK (pulse/step) pin.
+const uint16_t kDirPin = 12; ///< For the stepper driver DIR/CW (direction) pin.
+const uint16_t kEnaPin = 13; ///< For the stepper driver ENA/EN (enable) pin.
 /// @}
 
 /// @brief Serial properties.
@@ -34,12 +34,12 @@ const float kMicrostepAngle_degrees = kFullStepAngle_degrees / (kGearRatio * kMi
 /// @brief Motor states and targets.
 /// Distance.
 const float kDistance_degrees = 3600.0; ///< Target distance (degrees).
-uint64_t distance_microsteps = kDistance_degrees / kMicrostepAngle_degrees; ///< = 32000 Target distance (microsteps).
+uint32_t distance_microsteps = kDistance_degrees / kMicrostepAngle_degrees; ///< = 32000 Target distance (microsteps).
 /// Speed.
 const float kSpeed_RPM = 150.0; ///< Target speed (RPM).
 const float kSpeed_microsteps_per_s = (6.0 * kSpeed_RPM) / kMicrostepAngle_degrees; ///< = 8000.0 Target speed (microsteps/s).
-const uint64_t kMicrostepPeriod_us = 1000000.0 / kSpeed_microsteps_per_s; ///< = 125 Target speed based on the microstep period (us) between microsteps.
-uint64_t microstep_period_in_flux_us; ///< The microstep period (us) that is changing due to acceleration.
+const uint32_t kMicrostepPeriod_us = 1000000.0 / kSpeed_microsteps_per_s; ///< = 125 Target speed based on the microstep period (us) between microsteps.
+uint32_t microstep_period_in_flux_us; ///< The microstep period (us) that is changing due to acceleration.
 float p = 0.0; ///< ith speed (us), used to set microstep_period_in_flux_us.
 const float v0 = 0.0; ///< Base speed (microsteps/s) used to calculate the initial value of p. 
 const float f = 1000000.0; ///< Timer frequency (count of timer ticks per sec) (Hz).
@@ -48,7 +48,7 @@ uint64_t reference_time_us; ///< Reference time (us) for the microstep period.
 const float kAcceleration_microsteps_per_s_per_s = 3000.0; ///< Target acceleration (microsteps/s^2).
 const float R = kAcceleration_microsteps_per_s_per_s / (f * f); ///< Constant multiplier.
 const float m = -R; ///< Variable multiplier that depends on movement phase (m = -R for acceleration, m = 0 in-between, m = R for deceleration).
-int64_t i = 0; ///< Iteration counter.
+int32_t i = 0; ///< Iteration counter.
 float q = 0.0; ///< Variable to calculate a more accurate value of p at the expense of processing overhead (i.e., slower).
 /// @}
 
@@ -119,7 +119,7 @@ void CalculateNewSpeed() {
     p = p * (1 + q); // Equation 20.
     //p = p * (1 + q + (q * q)); // Equation 23.
     //p = p * (1 + q + (1.5 * q * q)); // Equation 22.
-    //Serial.print(F("p")); Serial.print(long(i)); Serial.print(F(" = ")); Serial.println(p);
+    //Serial.print(F("p")); Serial.print(i); Serial.print(F(" = ")); Serial.println(p);
     i++;
   }
   else {
